backend/requirements.txt
Flask
flask-jwt-extended
flask-bcrypt
flask_wtf
flask_cors
pytz
requests
pydantic
python-dotenv
psycopg2-binary
GitPython
google-auth-oauthlib
google-auth
google-api-core
google-generativeai
transformers
torch
